Products API

+++ buyersandsellers/vw_vendedor.py
@@ -2,6 +2,13 @@ from django.shortcuts import render
 from django.urls import reverse
 from django.http import HttpResponseRedirect
 from django.contrib.auth.models import Group, User
+from django.db.models import ProtectedError
+from django.conf import settings
+from random import randint
+from os.path import isfile
+from os import remove
+
+from operator import attrgetter
 
 from seguridad.mkitsafe import *
 from seguridad.models import Usr
@@ -15,10 +22,14 @@ def index( request ):
     if usuario.has_perm_or_has_perm_child( 'vendedor.agregar_vendedores_usuario' ):
         toolbar.append( { 'type' : 'link', 'view' : 'vendedor_nuevo', 'label' : '<i class="far fa-file"></i> Nuevo' } )
     data = []
-    desc =  usuario.descendencia()
-    for d in desc:
-        if d.groups.filter( name = 'Vendedor' ).exists():
-            data.append( Vendedor.objects.get( idusuario = d.pk ) )
+    if usuario.is_superuser \
+            or usuario.groups.filter( name__icontains = "Administrador" ).exists() \
+            or usuario.groups.filter( name__icontains = "Super-Administrador" ).exists():
+        data = Vendedor.get_Vendedores()
+    else:
+        gerente = Vendedor.get_from_usr( usuario )
+        if not gerente is None:
+            data = Vendedor.get_Vendedores( gerente )
     return render(
         request,
         'buyersandsellers/vendedor/index.html', {
@@ -31,40 +42,77 @@ def index( request ):
     
 @valida_acceso( [ 'vendedor.agregar_vendedores_usuario' ] )
 def new( request ):
+    usuario = Usr.objects.filter( id = request.user.pk )[ 0 ]
     if 'POST' == request.method:
-        frm = RegVendedorIn( request.POST, files = request.FILES )
+        if usuario.is_superuser \
+                or usuario.groups.filter( name__icontains = "Administrador" ).exists() \
+                or usuario.groups.filter( name__icontains = "Super-Administrador" ).exists():
+            frm = RegVendedorInAdmin( request.POST, files = request.FILES )    
+        else:
+            frm = RegVendedorIn( request.POST, files = request.FILES )
         if frm.is_valid():
-            perfil = Group.objects.get( name = 'Vendedor' )
+            perfil = Group.objects.get( name = '020 Vendedor' )
             usrdep = None
             obj = frm.save( commit = False )
             if not obj.reporta_a is None:
                 usrdep = Usr.objects.get( pk = obj.reporta_a.idusuario )
+            else:
+                usrdep = usuario
+                obj.reporta_a = Vendedor.get_from_usr( usuario )
             obj.username = obj.usuario
             obj.set_password( obj.contraseña )
             obj.save()
             obj.groups.add( perfil )
             obj.depende_de = usrdep
             obj.save()
+            upload_to = 'usuarios'
+            if "" != request.POST.get( 'token' ):
+                file = "{}{}/{}.imgx".format( settings.MEDIA_ROOT, upload_to, request.POST.get( 'token' ) )
+                if isfile( file ):
+                    f = open( file, "r" )
+                    if "r" == f.mode:
+                        img = f.read().strip()
+                        f.close()
+                        obj.fotografia = "{}/{}".format( upload_to, img )
+                        obj.save()
+                        remove( file )
             return HttpResponseRedirect( reverse( 'vendedor_ver', kwargs = { 'pk' : obj.pk } ) )
-    frm = RegVendedorIn( request.POST or None )
+    if usuario.is_superuser \
+            or usuario.groups.filter( name = "Administrador" ).exists() \
+            or usuario.groups.filter( name = "Super-Administrador" ).exists():
+        frm = RegVendedorInAdmin( request.POST or None )    
+    else:
+        frm = RegVendedorIn( request.POST or None )
     return render( request, 'global/form.html', {
-        'menu_main' : Usr.objects.filter( id = request.user.pk )[ 0 ].main_menu_struct(),
+        'menu_main' : usuario.main_menu_struct(),
         'footer' : True,
         'titulo' : 'Vendedores',
         'titulo_descripcion' : 'Nuevo',
-        'frm' : frm
+        'frm' : frm,
+        'uploader' : {
+            'url' : settings.UPLOADER_URL,
+            'site' : settings.UPLOADER_SITE,
+            'key' : settings.UPLOADER_KEY,
+            'onresponse' : '',
+            'type' : 'usuarios',
+            'excecute' : '',
+            'token' : randint( 1, 999999999 ),
+            'message' : "Archivo Cargado",
+        }
     } )
     
 @valida_acceso( [ 'vendedor.vendedores_usuario' ] )
 def see( request, pk ):
+    if not Vendedor.objects.filter( pk = pk ).exists():
+        return HttpResponseRedirect( reverse( 'seguridad_item_no_encontrado' ) )
     obj = Vendedor.objects.get( pk = pk )
-    frm = RegVendedor( instance = obj )
+    frm = RegVendedorAdmin( instance = obj )
     arbol = [ obj ] + obj.descendencia()
     for a in arbol:
         a = {
             'depth_name' : a.depth_name,
-            'cliente' : a.groups.filter( name = 'Cliente' ).exists(),
-            'vendedor' : a.groups.filter( name = 'Vendedor' ).exists()
+            'cliente' : a.groups.filter( name__icontains = 'Cliente' ).exists(),
+            'vendedor' : a.groups.filter( name__icontains = 'Vendedor' ).exists()
         }
     usuario = Usr.objects.filter( id = request.user.pk )[ 0 ]
     toolbar = []
@@ -88,18 +136,40 @@ def see( request, pk ):
     
 @valida_acceso( [ 'vendedor.actualizar_vendedores_usuario' ] )
 def update( request, pk ):
+    if not Vendedor.objects.filter( pk = pk ).exists():
+        return HttpResponseRedirect( reverse( 'seguridad_item_no_encontrado' ) )
+    usuario = Usr.objects.filter( id = request.user.pk )[ 0 ]
     obj = Vendedor.objects.get( pk = pk )
     if 'POST' == request.method:
-        frm = RegVendedor( instance = obj, data = request.POST, files = request.FILES )
+        if usuario.is_superuser \
+                or usuario.groups.filter( name__icontains = "Administrador" ).exists() \
+                or usuario.groups.filter( name__icontains = "Super-Administrador" ).exists():
+            frm = RegVendedorAdmin( instance = obj, data = request.POST, files = request.FILES )
+        else:
+            frm = RegVendedor( instance = obj, data = request.POST, files = request.FILES )
         if frm.is_valid():
             usrdep = None
             obj = frm.save( commit = False )
             if not obj.reporta_a is None:
                 usrdep = Usr.objects.get( pk = obj.reporta_a.idusuario )
+            else:
+                usrdep = usuario
+                obj.reporta_a = Vendedor.get_from_usr( usuario )
             obj.username = obj.usuario
             obj.save()
             obj.depende_de = usrdep
             obj.save()
+            upload_to = 'usuarios'
+            if "" != request.POST.get( 'token' ):
+                file = "{}{}/{}.imgx".format( settings.MEDIA_ROOT, upload_to, request.POST.get( 'token' ) )
+                if isfile( file ):
+                    f = open( file, "r" )
+                    if "r" == f.mode:
+                        img = f.read().strip()
+                        f.close()
+                        obj.fotografia = "{}/{}".format( upload_to, img )
+                        obj.save()
+                        remove( file )
             return HttpResponseRedirect( reverse( 'vendedor_ver', kwargs = { 'pk' : obj.pk } ) )
         else:
             return render( request, 'global/form.html', {
@@ -107,20 +177,49 @@ def update( request, pk ):
             'footer' : True,
             'titulo' : 'Vendedores',
             'titulo_descripcion' : obj,
-            'frm' : frm
+            'frm' : frm,
+            'uploader' : {
+                'url' : settings.UPLOADER_URL,
+                'site' : settings.UPLOADER_SITE,
+                'key' : settings.UPLOADER_KEY,
+                'onresponse' : '',
+                'type' : 'usuarios',
+                'excecute' : '',
+                'token' : randint( 1, 999999999 ),
+                'message' : "Archivo Cargado",
+            }
         } )
-    frm = RegVendedor( instance = obj )
+    if usuario.is_superuser \
+            or usuario.groups.filter( name__icontains = "Administrador" ).exists() \
+            or usuario.groups.filter( name__icontains = "Super-Administrador" ).exists():
+        frm = RegVendedorAdmin( instance = obj, data = request.POST or None )
+    else:
+        frm = RegVendedor( instance = obj, data = request.POST or None )
     return render( request, 'global/form.html', {
-        'menu_main' : Usr.objects.filter( id = request.user.pk )[ 0 ].main_menu_struct(),
+        'menu_main' : usuario.main_menu_struct(),
         'footer' : True,
         'titulo' : 'Vendedores',
         'titulo_descripcion' : obj,
-        'frm' : frm
+        'frm' : frm,
+        'uploader' : {
+            'url' : settings.UPLOADER_URL,
+            'site' : settings.UPLOADER_SITE,
+            'key' : settings.UPLOADER_KEY,
+            'onresponse' : '',
+            'type' : 'usuarios',
+            'excecute' : '',
+            'token' : randint( 1, 999999999 ),
+            'message' : "Archivo Cargado",
+        }
     } )
-
 
 @valida_acceso( [ 'vendedor.eliminar_vendedores_usuario' ] )
 def delete( request, pk ):
-    Vendedor.objects.get( pk = pk ).delete()
-    return HttpResponseRedirect( reverse( 'vendedor_inicio' ) )
+    try:
+        if not Vendedor.objects.filter( pk = pk ).exists():
+            return HttpResponseRedirect( reverse( 'seguridad_item_no_encontrado' ) )
+        Vendedor.objects.get( pk = pk ).delete()
+        return HttpResponseRedirect( reverse( 'vendedor_inicio' ) )
+    except ProtectedError:
+        return HttpResponseRedirect( reverse( 'seguridad_item_con_relaciones' ) )
